dbricks_helpers/demos/general/ingestion_frameworks/ingestion_framework_with_dabs/requirements.txt
# Core dependencies for Loyalty 2.0 DAB Framework
pyyaml>=6.0
databricks-cli>=0.18.0
click>=8.0

# Optional: For local testing
# pytest>=7.4.0
# black>=23.0.0
# ruff>=0.1.0


